+++ app/javascript/vue/tasks/nomenclature/new_taxon_name/components/Author/AuthorSource.vue
@@ -32,7 +32,7 @@
             :global-id="citation.global_id"
           />
         </p>
-        <div class="horizontal-left-content">
+        <div class="horizontal-left-content gap-small">
           <citation-pages
             @setPages="addPages"
             @save="triggerSave"